backend/requirements.txt
flask
flask-cors
Pillow
pytesseract
opencv-python-headless
numpy
python-dotenv
Werkzeug
gunicorn